refactor: extract extensions service
